app/requirements.txt
fastapi
uvicorn
sqlalchemy
reportlab
pdfrw
arabic-reshaper
python-bidi
bcrypt
python-jose
requests
jinja2
python-multipart
python-dotenv